some fixes

+++ vllm_ascend/worker/model_runner_v1.py
@@ -3389,8 +3389,7 @@ class NPUModelRunner(GPUModelRunner):
                             block_size=block_size,
                             num_kv_heads=1,
                             head_size=128,
-                            # TODO(cmq): REVERT ME after debugging
-                            dtype=torch.bfloat16,
+                            dtype=torch.int8,
                             compress_ratio=4,
                         ))
                 elif layer_id % 2 != 0:
